Adds deleted id as an arg to Store#destroyResource
Depending on how you're listening to events on a Store, you might want to listen for `destroy` events, but that event only contains the resources that are still left in the Store.  Adding an id argument onto the event allows for various interfaces to remove the item if they need it.